+++ platform/services/geti_controller/app/platform_operations/cluster.py
@@ -1,3 +1,6 @@
+# Copyright (C) 2022-2025 Intel Corporation
+# LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
+
 import logging
 
 from kubernetes import client, config
